[REFACTOR] Library에 있는 매력 포인트, 키워드 로직을 분리
